% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-subscripts.R
\name{collect_subscripts}
\alias{collect_subscripts}
\title{Standardize slicing indices}
\usage{
collect_subscripts(env = parent.frame(), column_transform = TRUE)
}
\arguments{
\item{i}{\verb{[vector]}

A vector of indices.}
}
\description{
\code{collect_subscripts()} standardizes slicing indices to make it easier to
develop \code{[} methods. It takes care of turning missing indices into explicit
\code{NULL}s, tells you how many indexing arguments were provided by the user,
and optionally standardizes \code{x[i]} into \code{x[,j]}.
}
\details{

}
