% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_functions.R
\name{fac_and_contr}
\alias{fac_and_contr}
\title{Create a factor and specify contrasts.}
\usage{
fac_and_contr(x, levels, contrasts, ordered = FALSE)
}
\arguments{
\item{x}{An object coercible to \code{\link[base]{factor}}.}

\item{levels}{A character vector of levels for the factor.}

\item{contrasts}{A matrix of \code{\link[stats]{contrasts}} for the factor.}

\item{ordered}{A logical indicating whether or not the factor is ordered
(default \code{FALSE}).}
}
\description{
The \code{fac_and_contr} function is a convenience function which coerces
\code{x} to a factor with specified \code{levels} and \code{contrasts}.
}
\seealso{
\code{\link{named_contr_sum}} (unordered factors) and
  \code{\link{scaled_contr_poly}} (ordered factors).
}

