% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_terms.R
\name{formula.standardized.terms}
\alias{formula.standardized.terms}
\title{S3 method for \code{\link[stats]{formula}} for class \code{standardized.terms}.}
\usage{
\method{formula}{standardized.terms}(x, ...)
}
\arguments{
\item{x, ...}{See \code{\link[stats]{formula}}.}
}
\description{
Calls \code{\link[stats]{formula}} and then ensures that the result has
the proper terms attributes.
}

