% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_by.R
\name{makepredictcall.scaledby}
\alias{makepredictcall.scaledby}
\title{S3 method for \code{\link[stats]{makepredictcall}} for class \code{scaledby}.}
\usage{
\method{makepredictcall}{scaledby}(var, call)
}
\arguments{
\item{var, call}{See \code{\link[stats]{makepredictcall}}.}
}
\description{
Allows \code{\link{scale_by}} to be used within a regression
\code{\link[stats]{formula}} and ensures that the \code{predvars} attribute
makes the correct call to \code{scale_by}.
}

