% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_terms.R
\name{model.frame.standardized.terms}
\alias{model.frame.standardized.terms}
\title{S3 method for \code{\link[stats]{model.frame}} for class \code{standardized.terms}.}
\usage{
\method{model.frame}{standardized.terms}(formula, data = NULL, ...)
}
\arguments{
\item{formula, data, ...}{See \code{\link[stats]{model.frame}}.}
}
\description{
S3 method for \code{\link[stats]{model.frame}} for class \code{standardized.terms}.
}

