% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_terms.R
\name{model.matrix.standardized.terms}
\alias{model.matrix.standardized.terms}
\title{S3 method for \code{\link[stats]{model.matrix}} for class \code{standardized.terms}.}
\usage{
\method{model.matrix}{standardized.terms}(object, data = environment(object),
  ...)
}
\arguments{
\item{object, data, ...}{See \code{\link[stats]{model.matrix}}.}
}
\description{
S3 method for \code{\link[stats]{model.matrix}} for class \code{standardized.terms}.
}

