% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized-class.R
\name{print.standardized}
\alias{print.standardized}
\title{S3 \code{print} method for class \code{\link[=standardized-class]{standardized}}.}
\usage{
\method{print}{standardized}(x, ...)
}
\arguments{
\item{x}{An object of class \code{standardized}.}

\item{...}{Not used.}
}
\description{
S3 \code{print} method for class \code{\link[=standardized-class]{standardized}}.
}

