% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{standardize.news}
\alias{standardize.news}
\title{Print the version history of the \code{standardize} package.}
\usage{
standardize.news()
}
\value{
The function prints the changes and new features for each version
  of the package (starting with the newest version).
}
\description{
Print the version history of the \code{standardize} package.
}
\examples{
standardize.news()

}
\author{
Christopher D. Eager <eager.stats@gmail.com>
}
