% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_terms.R
\name{standardize_terms}
\alias{standardize_terms}
\title{Standardize a formula for regression.}
\usage{
standardize_terms(formula, data, family = gaussian, scale = 1,
  na.action = "na.pass")
}
\arguments{
\item{formula}{A regression \code{\link[stats]{formula}}.}

\item{data}{A data.frame containing the variables in \code{formula}.}

\item{family}{A regression \code{\link[stats]{family}} (default gaussian).}

\item{scale}{The desired scale for the regression frame. Must be a single
positive number. See 'Details'.}

\item{na.action}{See \code{\link[stats]{model.frame}}.}
}
\value{
A terms object of class \code{standardized.terms}. The
  \code{predvars} attribute is altered to ensure that proper predictions
  are made when the object is used to create model frames and matrices.
  It also has a \code{standardized.scale} attribute which contains the
  \code{scale} argument passed to \code{standardize_terms}.
}
\description{
Create a terms object (of class \code{standardized.terms}) which places all
variables in \code{data} on the same scale, making regression output more
easy to interpret. For mixed effects regressions, this also offers
computational benefits, and for Bayesian regressions, it also makes
determining reasonable priors easier.
}
\details{
First \code{\link[stats]{model.frame}} is called. Then,
if \code{family = gaussian}, the response is checked to ensure that it is
numeric and has more than two unique values.  If \code{\link{scale_by}} is
used on the response in \code{formula}, then the \code{scale} argument to
\code{scale_by} is ignored and forced to \code{1}.  If \code{\link{scale_by}}
is not called, then \code{\link[base]{scale}} is used with default arguments.
For all other values for \code{family}, the response is not checked.

For the predictors in the formula, first any random effects grouping factor
in the formula is coerced to factor and unused levels are dropped.  The
levels of the resulting factor are then recorded in the \code{predvars}
attribute.  Then for the remaining predictors, regardless of their original
class, if they have only two unique non-\code{NA} values, they are coerced
to unordered factors.  Then, \code{\link{named_contr_sum}} and
\code{\link{scaled_contr_poly}} are called for unordered and ordered factors,
respectively, using the \code{scale} argument provided in the call
to \code{standardize_terms} as the \code{scale} argument to the contrast
functions.  For numeric variables, if the variable contains a call to
\code{\link{scale_by}}, then, regardless of whether the call to
\code{\link{scale_by}} specifies \code{scale}, the value of \code{scale}
in the call to \code{standardize_terms} is used.  If the numeric variable
does not contain a call to \code{\link{scale_by}}, then
\code{\link[base]{scale}} is called, ensuring that the result has
standard deviation \code{scale}.

With the default value of \code{scale = 1}, the result is a terms object
with a \code{predvars} attribute which ensures that a model frame created
using the terms object and the same data frame with which it was created
has numeric variables on unit scale, unordered factors with named sum
sum contrasts, and ordered factors with orthogonal polynomial contrasts
on unit scale.  For gaussian regressions, the response is also placed on
unit scale.  If \code{scale = 0.5}, then gaussian responses would still
be placed on unit scale, but unordered factors' named sum contrasts would
take on values {-0.5, 0, 0.5} rather than {-1, 0, 1}, the standard deviation
of each column in the contrast matrices for ordered factors would be
\code{0.5} rather than \code{1}, and the standard deviation of numeric
variables would be \code{0.5} rather than \code{1} (within-factor-level
in the case of \code{\link{scale_by}} calls).  The returned object
can be used as the \code{formula} argument to regression fitting functions
including those in the \code{lme4} and \code{rstanarm} packages, etc.
}
\section{Warning}{
 Offsets are not currently supported.
}
\examples{
# NEED TO ADD EXAMPLES

}
\seealso{
\code{\link[base]{scale}}, \code{\link{scale_by}},
  \code{\link{named_contr_sum}}, and \code{\link{scaled_contr_poly}}.
}

