% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_terms.R
\name{standardized_scale}
\alias{standardized_scale}
\title{Get the \code{scale} argument passed to \code{standardize}.}
\usage{
standardized_scale(object)
}
\arguments{
\item{object}{Any object for which a \code{\link[stats]{terms}} method
exists.}
}
\value{
A postive scalar corresponding to the \code{scale} argument passed
  to \code{\link{standardize_terms}} which created the
  \code{standardized.terms} object used in creating \code{object}.  If
  \code{standardize_terms} was not used, then \code{NULL} is returned.
}
\description{
Returns the \code{scale} argument from a \code{standardized.terms} object.
}

