% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized-class.R
\name{is.standardized}
\alias{is.standardized}
\title{Determine if an object has class \code{\link[=standardized-class]{standardized}}.}
\usage{
is.standardized(object)
}
\arguments{
\item{object}{An object of class \code{standardized}.}
}
\value{
\code{TRUE} if \code{object} is the result of a \code{\link{standardize}}
  call and \code{FALSE} otherwise.
}
\description{
Determine if an object has class \code{\link[=standardized-class]{standardized}}.
}

