% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-package.R
\docType{package}
\name{standardize-package}
\alias{standardize}
\alias{standardize-package}
\title{standardize: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/standardize}
  \item Report bugs at \url{https://github.com/DavisVaughan/standardize/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
