% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-package.R
\docType{package}
\name{standardize-package}
\alias{standardize}
\alias{standardize-package}
\title{standardize: Standardize Subscripts}
\description{
The goal of standardize is to ease the creation of data frame and
    array '[' methods by providing a way to standardize the subscripting
    arguments and provide useful information about the overall call, such
    as the number and type of arguments provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/standardize}
  \item Report bugs at \url{https://github.com/DavisVaughan/standardize/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
